Various beamforming fixes and improvements

+++ lib/buffers.h
@@ -152,7 +152,7 @@ int32_t get_buffer_data_ID(struct Buffer * buf, const int ID);
 
 uint32_t get_fpga_seq_num(struct Buffer * buf, const int ID);
 
-uint16_t get_streamID(struct Buffer * buf, const int ID);
+int32_t get_streamID(struct Buffer * buf, const int ID);
 
 struct timeval get_first_packet_recv_time(struct Buffer * buf, const int ID);
